+++ src/Bootstrap.php
@@ -126,8 +126,7 @@ class Bootstrap {
 		add_action(
 			'init',
 			function() {
-				( new Settings( self::$options, self::$config_path, $this->defined_constants ) )
-					->load_hooks();
+				( new Settings( self::$options, self::$config_path, $this->defined_constants ) )->load_hooks();
 			}
 		);
 		add_action(