+++ CHANGES.md
@@ -2,7 +2,7 @@
 * exit if called directly
 * exit during WP-CLI
 * no more testing `wp-config.php` everywhere
-* use `admin_init` hook
+* use `wp_loaded` hook for `process_filter_constants()`
 
 #### 2.7.2 / 2020-06-01
 * test `wp-config.php` everywhere, still occasional WSOD reports